refactor(update): extract Artwork dispatch to update/artwork.rs (commit 1/N)
Pull the 130-line Message::Artwork(...) inline match out of update/mod.rs
into a new file with a single dispatch_artwork method. The two API
fetchers (load_genre_album_ids, load_playlist_album_ids) move with it
since they're only consumed by the artwork pipeline.

The mod.rs match arm shrinks from a 130-line nested block to a 1-line
delegation:

    Message::Artwork(msg) => self.dispatch_artwork(msg),

Adding a new ArtworkMessage variant is now a 2-site edit
(app_message.rs + artwork.rs) instead of 3 (also touching mod.rs).
The handler functions themselves stay in albums.rs / collage.rs /
songs.rs unchanged.

No behavioral change. cargo build / clippy --all-targets -D warnings /
nightly fmt --check / test --workspace all clean (662 + 368 + 1 doctest).
